--- src/TestStoreContext/__tests__/index.test.js
@@ -1,8 +1,0 @@
-import {TestStoreContext} from ".."
-
-describe("TestStoreContext component", () => {
-  it('renders without crashing', () => {
-    const wrapper = shallow(<TestStoreContext store={{value: "test value"}}/>)
-    wrapper.unmount()
-  })
-})
