+++ src/components/NotFound/index.js
@@ -4,13 +4,13 @@ import {withTranslation} from 'react-i18next'
 
 import {Button, Typography} from '@material-ui/core'
 
-import {routes} from '../lib/router'
-import empty from "../assets/empty_state.svg"
+import {routes} from '../../lib/router'
+import empty from "../../assets/empty_state.svg"
 
 import "./not-found.sass"
 
-import {useDimensions} from '../hooks'
-import Centered from '../components/Centered'
+import {useDimensions} from '../../hooks'
+import Centered from '../Centered'
 
 const NotFound = ({t}) => {
